Phase A+B: enriched alerts + SQLite logging + GitHub data sync
Backtest validation
- run_per_ticker_backtest.py: incremental-save backtest runner that bypasses
  the buggy portfolio aggregation. Loops per ticker, persists after each so a
  crash on N+1 doesn't lose N's results.
- backtest_engine.py: portfolio_backtest int64-dtype fix (was crashing in
  pandas 2.x+ when assigning float returns to int Series).
- backtest_results_per_ticker.json + backtest_summary.txt: 4yr / 18-ticker
  results. 72% Sharpe-positive, median Sharpe 0.65, avg AnnRet +6.05%.

Per-asset threshold tuning (profiles.json)
- Loss-makers (Sharpe < 0): AMZN, GME, SCHD, ETH-USD, TLT → ±0.65
- Marginal (Sharpe 0–0.5): BTC-USD, AMD, MSFT → ±0.55
- Working (Sharpe ≥ 0.5): all others keep default ±0.45
- alert_system.evaluate now consults profile thresholds first, falls back
  to env defaults.

Phase A — enriched alert template (signal_enrichment.py)
- expected_move(): ATR-scaled 5d expected % and $ change. Honest vol-based
  estimate — magnitude, not direction prediction.
- suggest_option(): Polygon options chain → nearest weekly 7–14 DTE,
  1-strike OTM in signal direction. Skipped cleanly for crypto / no-options.
- get_asset_edge(): historical Sharpe / Win / AnnRet from backtest table.
- format_enrichment_block(): Telegram-Markdown formatter.
- alert_system.evaluate now builds enriched messages with price, expected
  move, historical edge, and suggested contract.

Phase B — SQLite + outcome scoring (signal_logger.py)
- Schema: scan_runs / scans / alerts / outcomes
- Every scan row persisted with full score breakdown + price
- Every alert persisted with enrichment snapshot for forensics
- score_pending_outcomes(): fetches +1d/+5d/+10d realised return per alert,
  flags correct_5d for win/loss tracking. The dataset for Phase C calibration.

Phase B — GitHub data sync (github_push.py)
- Clones repo to tmp, copies state/hype_fear.db → data/hype_fear.db,
  writes data/summary.json + data/daily_reports/YYYY-MM-DD.md, commits + pushes.
- Idempotent: no-op when no changes.
- Reuses the existing GITHUB_TOKEN pattern from Markets/github_push.py.

Worker loop integration (alert_system.run_loop)
- Daily (once per UTC day): score pending outcomes
- Daily after 03:00 UTC: push DB + summary + daily report to GitHub
- Per-scan: persist all rows to SQLite

Deploy config
- render.yaml: GITHUB_TOKEN added as a managed secret env var.
- .env.example: documents GITHUB_TOKEN with required permissions.
- .gitignore: keep backtest_results_per_ticker.json (loaded at runtime).

+++ backtest_engine.py
@@ -977,8 +977,10 @@ class EnhancedBacktester:
         weights = np.array([1.0 / num_assets] * num_assets)
 
         # Apply signals to weights (long only for positive signals)
-        portfolio_returns = pd.Series(0, index=combined.index)
-        previous_signals = pd.Series(0, index=signals.keys())  # Track previous positions for trade detection
+        # NB: dtype float64 — assigning float returns to an int64 Series raises
+        # TypeError on pandas 2.x+.
+        portfolio_returns = pd.Series(0.0, index=combined.index, dtype='float64')
+        previous_signals = pd.Series(0.0, index=signals.keys(), dtype='float64')  # Track previous positions for trade detection
 
         for date in combined.index:
             daily_signals = combined.loc[date]
